move stacking and alignment to separate module
